net, tests: Move shared bridge_nncp and NAD fixtures
Extract bridge_nncp, bridge_nad_a, and bridge_nad_b from
tests/network/l2_bridge/ into tests/fixtures/network/l2_bridge.py
so they are defined above the test suite that uses them.
To be used in upcoming metric test.

tests/network/l2_bridge/conftest.py and
tests/network/l2_bridge/nad_ref_change/conftest.py load the
fixtures via pytest_plugins.
